Implemented logout. added edit address & user details apis.

+++ MyntraClone/src/main/java/com/myntra/service/impl/RefreshTokenServiceImpl.java
@@ -55,6 +55,8 @@ public class RefreshTokenServiceImpl implements RefreshTokenService {
 
 	@Override
 	public void deleteToken(String input) {
+		if(refreshTokenRepository.findByToken(input).isPresent()) {
 		refreshTokenRepository.deleteById(input);
+		}
 	}
 }
